:bug: fix: 유저의 관심 콘서트 단건 삭제 API 에러 조건 추가 및 엔드포인트 변경

+++ src/user/user.service.ts
@@ -422,6 +422,14 @@ export class UserService {
   async removeInterestConcertById(userId: number, concertId: number) {
     await this.validateUser(userId);
 
+    const concert = await this.prismaService.concert.findUnique({
+      where: { id: concertId },
+    });
+
+    if (!concert) {
+      throw new NotFoundException(ErrorCode.CONCERT_NOT_FOUND);
+    }
+
     await this.prismaService.userInterestConcert.deleteMany({
       where: {
         userId,
